% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice_seeminglyunrelatedregression.R
\name{mice_seeminglyunrelatedregression}
\alias{mice_seeminglyunrelatedregression}
\title{Fit a seeminglyunrelated regression to mice data and get the coefficents and covariance matrix after applying Rubin's rules}
\usage{
mice_seeminglyunrelatedregression(
  imputed_data_,
  imputation_numb_,
  system_,
  numb_covar_,
  method_
)
}
\arguments{
\item{imputed_data_}{is your final imputed dataset given by the mice function
mice}

\item{imputation_numb_}{is the number of imputations conducted}

\item{system_}{is the list of regression equationa you want to fit}

\item{numb_covar_}{is the number of covariates you have in your statistical models}

\item{method_}{is the type of model you want systemfit to use. The options are the same
as the method option in systemfit.}
}
\value{
data_toreturn is a list, the first object is the coefficents of the regression
across the imputed datasets. The second object is the covariance matrix combined across the
imputed datasets
}
\description{
Fit a seeminglyunrelated regression to mice data and get the coefficents and covariance matrix after applying Rubin's rules
}
